(PC-2113): Add count_all_bookings() to booking repository
